requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
bootstrap4==0.1.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
cloudinary==1.26.0
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.3
django-cache-url==3.2.3
django-easy-pdf==0.1.1
django-embed-video==1.4.0
django-facebook==6.0.3
django-filter==2.4.0
django-heroku==0.3.1
django-installed-apps-command==2020.12.3
django-multiupload==0.5.2
django-search-views==0.3.7
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
django-yarn==0.0.2
django-yarnpkg==6.0.1
EasyProcess==0.3
environs==8.0.0
facebook-sdk==3.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
imgkit==1.2.2
iniconfig==1.1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
packaging==21.0
pdfkit==0.6.1
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
pytest==6.2.4
pytest-xvfb==2.0.0
python-bidi==0.4.2
python-decouple==3.4
python-dotenv==0.17.1
pytz==2021.1
PyVirtualDisplay==2.2
reportlab==3.5.68
requests==2.25.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
Unidecode==1.2.0
urllib3==1.26.4
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xvfbwrapper==0.2.9
